add egl

+++ examples/tests/DataToJsonFiles.cpp
@@ -8,9 +8,7 @@
 using json = nlohmann::json;
 TEST_CASE("DataToJsonFiles", "[DataToJsonFiles]")
 {
-//	std::string dir = "d:/output/";
-
-	for (int i = 0; i < pointsCN220_L.size(); ++i)
+	for (int i = 0; i < (int)pointsCN220_L.size(); ++i)
 	{
 		auto const &data = pointsCN220_L[i];
 		auto j = Common::dataToJson(data);
@@ -24,7 +22,7 @@ TEST_CASE("DataToJsonFiles", "[DataToJsonFiles]")
 		}
 		o << std::setw(4) << j;
 	}
-	for (int i = 0; i < pointsCN220_R.size(); ++i)
+	for (int i = 0; i < (int)pointsCN220_R.size(); ++i)
 	{
 		auto const &data = pointsCN220_R[i];
 		auto j = Common::dataToJson(data);
